src: Initial commit of lu_build_factors, lu_condest, lu_matrix_norm and lu_residual_test. Replacing lu slice fields with Vecs until istore/xstore pointer aliasing can be resolved. Resolving type related build errors, including indexing with lu_int.

+++ src/lu_def.rs
@@ -9,13 +9,13 @@ pub(crate) const FACTORIZE_BUMP: lu_int = 3;
 pub(crate) const BUILD_FACTORS: lu_int = 4;
 
 pub(crate) fn lu_iswap(x: &mut [lu_int], i: lu_int, j: lu_int) {
-    let t = x[i];
-    x[i] = x[j];
-    x[j] = t;
+    let t = x[i as usize];
+    x[i as usize] = x[j as usize];
+    x[j as usize] = t;
 }
 
 pub(crate) fn lu_fswap(x: &mut [f64], i: lu_int, j: lu_int) {
-    let t = x[i];
-    x[i] = x[j];
-    x[j] = t;
+    let t = x[i as usize];
+    x[i as usize] = x[j as usize];
+    x[j as usize] = t;
 }
